+++ django_comparison_dashboard/sources/core.py
@@ -2,10 +2,32 @@ import abc
 import logging
 
 import pandas as pd
+import pandera
+import pandera.io
 from django.shortcuts import get_object_or_404
-from frictionless import Resource, validate_resource
 
-from django_comparison_dashboard import models, settings
+from django_comparison_dashboard import forms, models, settings
+
+
+class SourceRegistry:
+    _instance = None
+    sources = {}
+
+    def __new__(cls):
+        if cls._instance is None:
+            cls._instance = super().__new__(cls)
+        return cls._instance
+
+    @classmethod
+    def register(cls, source: type["DataSource"]):
+        cls.sources[source.name] = source
+
+    def __getitem__(self, item):
+        if item in self.sources:
+            return self.sources[item]
+        raise KeyError(
+            f"Can not find source '{item}' in source registry. Maybe you forgot to register it in the first place?"
+        )
 
 
 class ScenarioValidationError(Exception):
@@ -13,15 +35,21 @@ class ScenarioValidationError(Exception):
 
 
 class Scenario(abc.ABC):
-    source: "DataSource" = None
+    source_name: str = None
 
-    def __init__(self, scenario_id, data_type: settings.DataType):
+    def __init__(self, scenario_id, data_type: settings.DataType | str):
+        if self.source_name is None:
+            raise RuntimeError("Source name not defined.")
         self.id = scenario_id
-        self.data_type = data_type
+        self.data_type = data_type if isinstance(data_type, settings.DataType) else settings.DataType[data_type]
 
-    @abc.abstractmethod
     def __str__(self):
         """Return string representation of scenario"""
+        return self.id
+
+    @property
+    def source(self):
+        return SourceRegistry()[self.source_name]
 
     def is_present(self) -> bool:
         return models.Scenario.objects.filter(name=self.id, source__name=self.source.name).exists()
@@ -36,7 +64,7 @@ class Scenario(abc.ABC):
         self._store_in_db(data)
         logging.info(f"Successfully downloaded scenario '{self}'.")
 
-    def _store_in_db(self, data: list[dict] | pd.DataFrame):
+    def _store_in_db(self, data: pd.DataFrame):
         """
         Store data into corresponding database model (scalar or timeseries)
 
@@ -47,12 +75,19 @@ class Scenario(abc.ABC):
         """
         source = models.Source.objects.get_or_create(name=self.source.name)[0]
         scenario = models.Scenario.objects.get_or_create(name=self.id, source=source)[0]
-        data_model = models.ScalarData if self.data_type == settings.DataType.Scalar else models.TimeseriesData
-        data_model.objects.bulk_create(data_model(scenario=scenario, **item) for item in data)
+        if self.data_type == settings.DataType.Scalar:
+            data_model = models.ScalarData
+        elif self.data_type == models.TimeseriesData:
+            data_model = models.TimeseriesData
+        else:
+            raise TypeError(f"Unknown data type '{self.data_type}'.")
+        data_model.objects.bulk_create(
+            data_model(scenario=scenario, **item) for item in data.to_dict(orient="records")
+        )
 
-    def _validate(self, data: dict | pd.DataFrame) -> None:
+    def _validate(self, data: pd.DataFrame) -> None:
         """
-        Validate given data using frictionless and source-related schema
+        Validate given data using pandera and source-related schema
 
         Parameters
         ----------
@@ -62,32 +97,24 @@ class Scenario(abc.ABC):
         Returns
         -------
         dict
-            in case of an error a frictionless error report is returned, otherwise None is returned
+            in case of an error a report is returned, otherwise None is returned
 
         Raises
         ------
-        ScenarioValidationError
+        pandera.errors.SchemaError
             if scenario data does not fit into OEDatamodel format
         """
         logging.info(f"Validating data for scenario {self}...")
-        resource = Resource(
-            name=str(self.data_type),
-            profile="tabular-data-resource",
-            data=data,
-            schema=settings.MODEX_OUTPUT_SCHEMA[str(self.data_type)],
-        )
-        report = validate_resource(resource)
-        if report["stats"]["errors"] != 0:
-            error = ScenarioValidationError(f"Could not validate scenario data for scenario '{self}'.")
-            error.add_note(str(report))
-            raise error
+        schema = pandera.io.from_frictionless_schema(settings.MODEX_OUTPUT_SCHEMA[str(self.data_type)])
+        schema.validate(data, lazy=True)
 
 
 class DataSource(abc.ABC):
     name: str = None
+    scenario: type[Scenario] = None
+    form = forms.DataSourceUploadForm
 
     @classmethod
-    @abc.abstractmethod
     def list_scenarios(cls) -> list[Scenario]:
         """
         List all available scenarios