+++ django_comparison_dashboard/urls.py
@@ -8,4 +8,12 @@ urlpatterns = [
     path("index/", views.index, name="index"),
     path("filters/", views.get_filters, name="filters"),
     path("scalars/", views.plot_scalar_data),
+    path("", views.IndexView.as_view(), name="index"),
+    path("dashboard/", views.DashboardView.as_view(), name="dashboard"),
+    path("scalars/", views.scalar_data_plot),
+    path("scalars/data/", views.scalar_data_table),
+    path("scenarios/", views.ScenarioSelectionView.as_view(), name="scenarios"),
+    path("scenario_detail/", views.ScenarioDetailView.as_view(), name="scenario_detail"),
+    path("upload/", views.UploadView.as_view(), name="upload"),
+    path("scenario_form/", views.ScenarioFormView.as_view(), name="scenario_form"),
 ]