+++ CHANGELOG.md
@@ -6,6 +6,9 @@ and this project tries to adhere to [Semantic Versioning](https://semver.org/spe
 
 ## [Unreleased]
 ### Added
+- index view
+- scenario selection view
+- upload view for scenario data uploads
 - models for scenario data and sources
 - MODEX source
 - downloading of MODEX scenario data